refactor(swap): replace custom Decimal class with bigint planck values
Remove the swaps-port Decimal class (128 LOC) and replace all usage
across 20 files with raw bigint planck values + @talismn/util utilities.

Key changes:
- QuoteParams.fromAmount and ExchangeParams.fromAmount: Decimal → bigint
- useSwapQuoteManager toAmount: Decimal | null → bigint | null
- useFastBalance/useSubstrateBalance return bigint instead of Decimal
- TokenAmountInput props all use bigint
- useFiatValueForAmount accepts planck: bigint instead of amount: Decimal
- Swap modules (simpleswap, stealthex, lifi) use planckToTokens for
  human-readable amounts and parseUserInputToPlanck for API responses
- New swap-utils.ts with parseUserInputToPlanck and
  parseUserInputToPlanckOrUndefined utilities
- Delete swaps-port/Decimal.ts

+++ apps/extension/src/ui/domains/Swap/hooks/useSwapErc20Approval.ts
@@ -14,7 +14,6 @@ import type {
 } from "../swap-modules/common.swap-module"
 import type { swapModules } from "../swaps.api"
 import { allEvmChains } from "../swaps-port/allEvmChains"
-import type { Decimal } from "../swaps-port/Decimal"
 
 /**
  * Manages ERC20 approval state for the selected swap module.
@@ -24,7 +23,7 @@ export const useSwapErc20Approval = (params: {
   selectedModule: (typeof swapModules)[number] | undefined
   fromAsset: SwappableAssetWithDecimals | null
   toAsset: SwappableAssetWithDecimals | null
-  fromAmount: Decimal
+  fromAmount: bigint
   fromAddress: string | null
   toAddress: string | null
   selectedSubProtocol: string | undefined
